+++ src/index.stories.tsx
@@ -34,19 +34,20 @@ Anonymous.args = {
   baseURL: 'https://www.aisuru.com',
   uiLang: 'IT',
   spokenLang: 'IT',
-  layout: 'ZOOMED_FULL_BODY',
-  showInstruct: 'false',
-  showSettings: 'true',
-  showClear: 'false',
-  showAIicon: 'true',
-  showWhyThisAnswer: 'true',
-  showTypingText: 'false',
-  showOnlyLastMessages: 'false',
-  showTranslationOriginal: 'false',
-  showCopyButton: 'false',
-  showShare: 'true',
-  showLogin: 'false',
-  enableAudio: 'true',
+  enableAudio: true,
+};
+
+export const WithPreviousSession = Template.bind({});
+WithPreviousSession.args = {
+  ownerUserName: 'nzambello',
+  memoriName: 'Nicola',
+  tenantID: 'www.aisuru.com',
+  engineURL: 'https://engine.memori.ai',
+  apiURL: 'https://backend.memori.ai',
+  baseURL: 'https://www.aisuru.com',
+  sessionID: '1234567890',
+  uiLang: 'IT',
+  spokenLang: 'IT',
 };
 
 export const Nunzio = Template.bind({});
@@ -54,6 +55,7 @@ Nunzio.args = {
   ownerUserName: 'nunzio.fiore',
   memoriName: 'Nunzio',
   tenantID: 'aisuru.com',
+  engineURL: 'https://engine.memori.ai',
   apiURL: 'https://backend.memori.ai',
   baseURL: 'https://www.aisuru.com',
   uiLang: 'it',
@@ -69,6 +71,7 @@ Localhost.args = {
   ownerUserID: '04a8cff9-13d6-4367-9cb2-72b9af9ee494',
   tenantID: 'www.aisuru.com',
   apiURL: 'http://localhost:7778',
+  engineURL: 'http://localhost:7778',
   baseURL: 'http://localhost:3000',
   uiLang: 'EN',
   lang: 'IT',
@@ -84,6 +87,7 @@ LocalhostBoE.args = {
   memoriID: '2b094cd6-77b8-4e09-a807-0039b84c988e',
   ownerUserID: '04a8cff9-13d6-4367-9cb2-72b9af9ee494',
   tenantID: 'www.aisuru.com',
+  engineURL: 'http://localhost:7777',
   apiURL: 'http://localhost:7778',
   baseURL: 'http://localhost:3000',
   uiLang: 'EN',
@@ -123,8 +127,32 @@ Giovanna.args = {
   layout: 'ZOOMED_FULL_BODY',
 };
 
-export const GiovannaProva = Template.bind({});
-GiovannaProva.args = {
+export const GiovannaRPMProva = Template.bind({});
+GiovannaRPMProva.args = {
+  memoriName: 'Giovanna Test',
+  ownerUserName: 'andrea.patini3',
+  memoriID: '431d9819-c958-442c-a799-f90617371c0c',
+  ownerUserID: '58770358-a5db-4b49-b3a4-734fc468e745',
+  tenantID: 'aisuru-staging.aclambda.online',
+  engineURL: 'https://engine-staging.memori.ai',
+  apiURL: 'https://backend-staging.memori.ai',
+  baseURL: 'https://aisuru-staging.aclambda.online',
+  uiLang: 'EN',
+  spokenLang: 'IT',
+  layout: 'ZOOMED_FULL_BODY',
+  integrationID: '061898c0-2138-49da-a6b3-5d86267aad05',
+  initialQuestion: 'inizio simulazione',
+};
+
+export const GiovannaGLBProva = Template.bind({});
+GiovannaGLBProva.args = {
+  ...GiovannaRPMProva.args,
+  integrationID: 'd2099459-0f10-40cd-85e1-06e77a678723',
+  layout: 'ZOOMED_FULL_BODY',
+};
+
+export const GiovannaProvaWithPreviousSession = Template.bind({});
+GiovannaProvaWithPreviousSession.args = {
   memoriName: 'Giovanna Test',
   ownerUserName: 'andrea.patini3',
   memoriID: '431d9819-c958-442c-a799-f90617371c0c',
@@ -138,6 +166,7 @@ GiovannaProva.args = {
   layout: 'ZOOMED_FULL_BODY',
   integrationID: 'e92ac275-39b5-474d-8f9e-826cc5284f1e',
   initialQuestion: 'inizio simulazione',
+  sessionID: 'c1d13cec-1571-41a0-90bc-15bd2d62e263',
 };
 
 export const CreaRiassunti = Template.bind({});
@@ -232,6 +261,7 @@ WithBatch.args = {
   memoriName: 'Nicola',
   tenantID: 'aisuru.com',
   apiURL: 'https://backend.memori.ai',
+  engineURL: 'https://engine.memori.ai',
   baseURL: 'https://www.aisuru.com',
   uiLang: 'it',
   showShare: true,
@@ -244,6 +274,7 @@ WithCustomUserAvatar.args = {
   memoriName: 'Nicola',
   tenantID: 'aisuru.com',
   apiURL: 'https://backend.memori.ai',
+  engineURL: 'https://engine.memori.ai',
   baseURL: 'https://www.aisuru.com',
   uiLang: 'it',
   showShare: true,