+++ src/index.tsx
@@ -66,6 +66,7 @@ export interface Props {
   customMediaRenderer?: WidgetProps['customMediaRenderer'];
   additionalSettings?: WidgetProps['additionalSettings'];
   userAvatar?: WidgetProps['userAvatar'];
+  useMathFormatting?: boolean;
 }
 
 const getPreferredLanguages = () => {
@@ -130,6 +131,7 @@ const Memori: React.FC<Props> = ({
   customMediaRenderer,
   additionalSettings,
   userAvatar,
+  useMathFormatting = false,
 }) => {
   const [memori, setMemori] = useState<IMemori>();
   const [tenant, setTenant] = useState<Tenant>();
@@ -225,10 +227,6 @@ const Memori: React.FC<Props> = ({
     }
   }, [uiLang]);
 
-  useEffect(() => {
-    installMathJax();
-  }, []);
-
   return (
     <I18nWrapper>
       <VisemeProvider>
@@ -290,6 +288,7 @@ const Memori: React.FC<Props> = ({
             customMediaRenderer={customMediaRenderer}
             additionalSettings={additionalSettings}
             userAvatar={userAvatar}
+            useMathFormatting={useMathFormatting}
             {...(tag && pin ? { personification: { tag, pin } } : {})}
           />
         ) : (
@@ -380,6 +379,7 @@ Memori.propTypes = {
   customMediaRenderer: PropTypes.func,
   additionalSettings: PropTypes.any,
   userAvatar: PropTypes.oneOfType([PropTypes.string, PropTypes.any]),
+  useMathFormatting: PropTypes.bool,
 };
 
 export default Memori;